Added delete user UI

+++ HelpdeskSystem/Helpdesk.Services/UsersFacade.cs
@@ -219,13 +219,22 @@ namespace Helpdesk.Services
         /// </summary>
         /// <param name="id">The id of the user to be deleted</param>
         /// <returns>A response that indicates whether or not the deletion was successful</returns>
-        public DeleteUserResponse DeleteUser(int id)
+        public DeleteUserResponse DeleteUser(int id, string currentUser)
         {
             var response = new DeleteUserResponse();
 
             try
             {
                 var dataLayer = new UsersDataLayer();
+
+                UserDTO user = dataLayer.GetUser(id);
+
+                if (user.Username == currentUser)
+                {
+                    response.Status = HttpStatusCode.Forbidden;
+                    return response;
+                }
+
                 bool result = dataLayer.DeleteUser(id);
 
                 if (result)
